Små justeringer etter tilbakemeldinger
